requirements.txt
google-api-python-client
pandas
fastapi
uvicorn
nltk
vaderSentiment
transformers
torch
pydantic
python-dotenv
gunicorn
